Add JavaScript script to toggle class of header element
This commit adds a JavaScript script that toggles the class of the <header> element when the user clicks on the <div> element with the id 'toggle_header'. The script utilizes the jQuery API to achieve this functionality. The <header> element must always have one class: 'red' or 'green', never both at the same time and never empty. If the current class is 'red', clicking on 'toggle_header' updates the class to 'green', and vice versa.