Add JavaScript script to add <li> element to list
This commit adds a JavaScript script that adds a new <li> element with the text Item to a list when the user clicks on the <div> element with the id 'add_item'. The new element is appended to the <ul> element with the class 'my_list'. The script utilizes the jQuery API to achieve this functionality.